requirements.txt
llvmlite==0.41.0
mypy==1.6.0
mypy-extensions==1.0.0
numba==0.58.0
numpy==1.25.2
Pillow==10.0.1
tomli==2.0.1
typing_extensions==4.8.0
